feat: doc improvements, more tips
