+++ src/documentation/wiki-interface.ts
@@ -1,11 +1,11 @@
 import { RShell } from '../r-bridge/shell';
-import { FlowrGithubBaseRef, FlowrNpmRef, FlowrWikiBaseRef, getFileContentFromRoot } from './doc-util/doc-files';
+import { FlowrGithubBaseRef, FlowrWikiBaseRef, getFileContentFromRoot } from './doc-util/doc-files';
 import { getCliLongOptionOf, getReplCommand, multipleCliOptions } from './doc-util/doc-cli-option';
 import { printServerMessages } from './doc-util/doc-server-message';
 import { documentAllServerMessages } from './data/server/doc-data-server-messages';
-import { codeBlock, codeInline } from './doc-util/doc-code';
+import { codeBlock } from './doc-util/doc-code';
 import type { FileAnalysisRequestMessage } from '../cli/repl/server/messages/message-analysis';
-import { fileProtocol, removeRQuotes, requestFromInput } from '../r-bridge/retriever';
+import { fileProtocol, removeRQuotes } from '../r-bridge/retriever';
 import { DockerName } from './doc-util/doc-docker';
 import { documentReplSession, printReplHelpAsMarkdownTable } from './doc-util/doc-repl';
 import { printDfGraphForCode } from './doc-util/doc-dfg';
@@ -14,15 +14,13 @@ import { describeSchema } from '../util/schema';
 import { markdownFormatter } from '../util/text/ansi';
 import { defaultConfigFile } from '../cli/flowr-main-options';
 import { NewIssueUrl } from './doc-util/doc-issue';
-import { PipelineExecutor } from '../core/pipeline-executor';
-import { block, details } from './doc-util/doc-structure';
-import { TreeSitterExecutor } from '../r-bridge/lang-4.x/tree-sitter/tree-sitter-executor';
-import { FlowrAnalyzer } from '../project/flowr-analyzer';
+import { block } from './doc-util/doc-structure';
 import type { DocMakerArgs } from './wiki-mk/doc-maker';
 import { DocMaker } from './wiki-mk/doc-maker';
 import type { KnownParser } from '../r-bridge/parser';
 import type { GeneralDocContext } from './wiki-mk/doc-context';
 import { BuiltInProcName } from '../dataflow/environments/built-in';
+import { explainWritingCode } from './data/interface/doc-writing-code';
 
 async function explainServer(parser: KnownParser): Promise<string> {
 	documentAllServerMessages();
@@ -320,168 +318,6 @@ ${describeSchema(flowrConfigFileSchema, markdownFormatter)}
 	`;
 }
 
-function explainWritingCode(shell: RShell, ctx: GeneralDocContext): string {
-	return `
-_flowR_ can be used as a [module](${FlowrNpmRef}) and offers several main classes and interfaces that are interesting for extension writers 
-(see the [Visual Studio Code extension](${FlowrGithubBaseRef}/vscode-flowr) or the [core](${FlowrWikiBaseRef}/Core) wiki page for more information).
-
-### Using the ${ctx.link(RShell)} to Interact with R
-
-The ${ctx.link(RShell.name)} class allows interfacing with the \`R\`&nbsp;ecosystem installed on the host system.
-Please have a look at [flowR's engines](${FlowrWikiBaseRef}/Engines) for more information on alterantives (for example, the ${ctx.link(TreeSitterExecutor)}).
-
-${
-	block({
-		type:    'IMPORTANT',
-		content: `
-Each ${ctx.link(RShell.name)} controls a new instance of the R&nbsp;interpreter, 
-make sure to call ${codeInline(ctx.linkM(RShell, 'close', { codeFont: false, realNameWrapper: 'i' }) + '()')} when you are done.`
-	})
-}
-
-You can start a new "session" simply by constructing a new object with ${codeInline('new ' + ctx.link(RShell, { codeFont: false }) + '()')}.
-
-However, there are several options that may be of interest 
-(e.g., to automatically revive the shell in case of errors or to control the name location of the R process on the system).
-
-With a shell object (let's call it \`shell\`), you can execute R code by using ${ctx.linkM(RShell, 'sendCommand', { realNameWrapper: 'i' })}, 
-for example ${codeInline('shell.' + ctx.linkM(RShell, 'sendCommand', { codeFont: false, hideClass: true }) + '("1 + 1")')}. 
-However, this does not return anything, so if you want to collect the output of your command, use
-${ctx.linkM(RShell, 'sendCommandWithOutput', { realNameWrapper: 'i' })} instead.
-
-Besides that, the command ${ctx.linkM(RShell, 'tryToInjectHomeLibPath')} may be of interest, as it enables all libraries available on the host system.
-
-### Creating _flowR_ analyses
-
-Nowadays, instances of ${ctx.link(FlowrAnalyzer)} should be used as central frontend to get analysis results from _flowR_.
-For example, a program slice can be created like this:
-
-${
-	codeBlock('ts', `
-const analyzer = await new FlowrAnalyzerBuilder(requestFromInput('x <- 1\\ny <- x\\nx')).build();
-const result = await analyzer.query([
-	{
-		type:     'static-slice',
-		criteria: ['3@x']
-	}
-]);
-//console.log(result['static-slice']);
-`)
-}
-        
-### The Pipeline Executor
-
-Once, in the beginning, _flowR_ was meant to produce a dataflow graph merely to provide *program slices*. 
-However, with continuous updates, the [dataflow graph](${FlowrWikiBaseRef}/Dataflow-Graph) repeatedly proves to be the more interesting part.
-With this, we restructured _flowR_'s originally *hardcoded* pipeline to be far more flexible. 
-Now, it can be theoretically extended or replaced with arbitrary steps, optional steps, and what we call 'decorations' of these steps. 
-In short, a slicing pipeline using the ${ctx.link(PipelineExecutor)} looks like this:
-
-${
-	codeBlock('ts', `
-const slicer = new ${PipelineExecutor.name}(DEFAULT_SLICING_PIPELINE, {
-  parser:    new ${RShell.name}(),
-  request:   ${requestFromInput.name}('x <- 1\\nx + 1'),
-  criterion: ['2@x']
-})
-const slice = await slicer.allRemainingSteps()
-// console.log(slice.reconstruct.code)
-`)
-}
-
-${
-	details('More Information', `
-
-If you compare this, with what you would have done with the old (and removed) \`SteppingSlicer\`, 
-this essentially just requires you to replace the \`SteppingSlicer\` with the ${ctx.link(PipelineExecutor)}
-and to pass the ${ctx.link('DEFAULT_SLICING_PIPELINE')} as the first argument.
-The ${ctx.link(PipelineExecutor)}...
-
-1. Provides structures to investigate the results of all intermediate steps
-2. Can be executed step-by-step
-3. Can repeat steps (e.g., to calculate multiple slices on the same input)
-
-See the in-code documentation for more information.
-
-	`)
-}
-
-
-### Generate Statistics
-
-
-<details>
-
-<summary>Adding a New Feature to Extract</summary>
-
-In this example, we construct a new feature to extract, with the name "*example*".
-Whenever this name appears, you may substitute this with whatever name fits your feature best (as long as the name is unique).
-
-1. **Create a new file in \`src/statistics/features/supported\`**\\
-   Create the file \`example.ts\`, and add its export to the \`index.ts\` file in the same directory (if not done automatically).
-
-2. **Create the basic structure**\\
-   To get a better feel of what a feature must have, let's look
-   at the basic structure (of course, due to TypeScript syntax,
-   there are other ways to achieve the same goal):
-
-   \`\`\`ts
-   const initialExampleInfo = {
-       /* whatever start value is good for you */
-       someCounter: 0
-   }
-
-   export type ExampleInfo = Writable<typeof initialExampleInfo>
-
-   export const example: Feature<ExampleInfo> = {
-    name:        'Example Feature',
-    description: 'A longer example description',
-
-    process(existing: ExampleInfo, input: FeatureProcessorInput): ExampleInfo {
-      /* perform analysis on the input */
-      return existing
-    },
-
-    initialValue: initialExampleInfo
-   }
-   \`\`\`
-
-   The \`initialExampleInfo\` type holds the initial values for each counter that you want to maintain during the feature extraction (they will usually be initialized with 0). The resulting \`ExampleInfo\` type holds the structure of the data that is to be counted. Due to the vast amount of data processed, information like the name and location of a function call is not stored here, but instead written to disk (see below).
-
-   Every new feature must be of the \`Feature<Info>\` type, with \`Info\` referring to a \`FeatureInfo\` (like \`ExampleInfo\` in this example). Next to a \`name\` and a \`description\`, each Feature must provide:
-
-   - a processor that extracts the information from the input, adding it to the existing information.
-   - a function returning the initial value of the information (in this case, \`initialExampleInfo\`).
-
-3. **Add it to the feature-mapping**\\
-   Now, in the \`feature.ts\` file in \`src/statistics/features\`, add your feature to the \`ALL_FEATURES\` object.
-
-Now, we want to extract something. For the *example* feature created in the previous steps, we choose to count the amount of \`COMMENT\` tokens.
-So we define a corresponding [XPath](https://developer.mozilla.org/en-US/docs/Web/XPath) query:
-
-\`\`\`ts
-const commentQuery: Query = xpath.parse('//COMMENT')
-\`\`\`
-
-Within our feature's \`process\` function, running the query is as simple as:
-
-\`\`\`ts
-const comments = commentQuery.select({ node: input.parsedRAst })
-\`\`\`
-
-Now we could do a lot of further processing, but for simplicity, we only record every comment found this way:
-
-\`\`\`ts
-appendStatisticsFile(example.name, 'comments', comments, input.filepath)
-\`\`\`
-
-We use \`example.name\` to avoid duplication with the name that we’ve assigned to the feature. It corresponds to the name of the folder in the statistics output.
-\`'comments'\` refers to a freely chosen (but unique) name, that will be used as the name for the output file within the folder. The \`comments\` variable holds the result of the query, which is an array of nodes. Finally, we pass the \`filepath\` of the file that was analyzed (if known), so that it can be added to the statistics file (as additional information).
-
-</details>
-	`;
-}
-
 /**
  * https://github.com/flowr-analysis/flowr/wiki/Interface
  */
@@ -497,19 +333,14 @@ this wiki page explains how to interface with _flowR_ in more detail.
 In general, command line arguments and other options provide short descriptions on hover over.
 
 * [💻 Using the REPL](#using-the-repl)
-* [💬 Communicating with the Server](#communicating-with-the-server)
 * [⚙️ Configuring FlowR](#configuring-flowr)
 * [⚒️ Writing Code](#writing-code)
+* [💬 Communicating with the Server](#communicating-with-the-server)
 
 <a id='using-the-repl'></a>
 ## 💻 Using the REPL
 
 ${await explainRepl(treeSitter, ctx)}
-
-<a id='communicating-with-the-server'></a>
-## 💬 Communicating with the Server
-
-${await explainServer(shell)}
 
 
 <a id='configuring-flowr'></a>
@@ -521,6 +352,11 @@ ${explainConfigFile()}
 ## ⚒️ Writing Code
 
 ${explainWritingCode(shell, ctx)}
+
+<a id='communicating-with-the-server'></a>
+## 💬 Communicating with the Server
+
+${await explainServer(shell)}
 `;
 	}
 }
