+++ wiki/Analyzer.md
@@ -1,4 +1,4 @@
-_This document was generated from '[src/documentation/wiki-analyzer.ts](https://github.com/flowr-analysis/flowr/tree/main//src/documentation/wiki-analyzer.ts)' on 2026-02-01, 18:35:49 UTC presenting an overview of flowR's analyzer (v2.8.15, using R v4.5.2). Please do not edit this file/wiki page directly._
+_This document was generated from '[src/documentation/wiki-analyzer.ts](https://github.com/flowr-analysis/flowr/tree/main//src/documentation/wiki-analyzer.ts)' on 2026-02-01, 18:58:57 UTC presenting an overview of flowR's analyzer (v2.8.15, using R v4.5.2). Please do not edit this file/wiki page directly._
 
 - [Overview](#Overview)
   - [Overview of the Analyzer](#Overview_of_the_Analyzer)  
@@ -97,7 +97,7 @@ Likewise, <a href="https://github.com/flowr-analysis/flowr/tree/main//src/projec
 Again, <a href="https://github.com/flowr-analysis/flowr/tree/main//src/project/flowr-analyzer.ts#L245"><code>FlowrAnalyzer::<i>peekDataflow</i></code></a> allows you to inspect the dataflow graph if it was already computed (but without triggering a computation).
 * <a href="https://github.com/flowr-analysis/flowr/tree/main//src/project/flowr-analyzer.ts#L254"><code>FlowrAnalyzer::<b>controlflow</b></code></a> to compute the [Control Flow Graph](https://github.com/flowr-analysis/flowr/wiki/Control%20Flow%20Graph)\
 Also, <a href="https://github.com/flowr-analysis/flowr/tree/main//src/project/flowr-analyzer.ts#L258"><code>FlowrAnalyzer::<i>peekControlflow</i></code></a> returns the control flow graph if it was already computed but without triggering a computation.
-* <a href="https://github.com/flowr-analysis/flowr/tree/main//src/project/flowr-analyzer.ts#L262"><code>FlowrAnalyzer::<b>callGraph</b></code></a> to compute the [call graph](https://github.com/flowr-analysis/flowr/wiki/dataflow-graph#perspectives-cg) of the analyzed code\
+* <a href="https://github.com/flowr-analysis/flowr/tree/main//src/project/flowr-analyzer.ts#L262"><code>FlowrAnalyzer::<b>callGraph</b></code></a> to compute the [call graph](https://github.com/flowr-analysis/flowr/wiki/wiki/dataflow-graph#perspectives-cg) of the analyzed code\
 Likewise, <a href="https://github.com/flowr-analysis/flowr/tree/main//src/project/flowr-analyzer.ts#L266"><code>FlowrAnalyzer::<i>peekCallGraph</i></code></a> allows you to inspect the call graph if it was already computed (but without triggering a computation).
 * <a href="https://github.com/flowr-analysis/flowr/tree/main//src/project/flowr-analyzer.ts#L270"><code>FlowrAnalyzer::<b>query</b></code></a> to run [queries](https://github.com/flowr-analysis/flowr/wiki/Query-API) on the analyzed code.
 * <a href="https://github.com/flowr-analysis/flowr/tree/main//src/project/flowr-analyzer.ts#L276"><code>FlowrAnalyzer::<b>runSearch</b></code></a> to run a search query on the analyzed code using the [search API](https://github.com/flowr-analysis/flowr/wiki/Search%20API)
