+++ wiki/Normalized AST.md
@@ -1,4 +1,4 @@
-_This document was generated from '[src/documentation/wiki-normalized-ast.ts](https://github.com/flowr-analysis/flowr/tree/main//src/documentation/wiki-normalized-ast.ts)' on 2026-02-01, 18:35:49 UTC presenting an overview of flowR's normalized ast (v2.8.15, using R v4.5.2). Please do not edit this file/wiki page directly._
+_This document was generated from '[src/documentation/wiki-normalized-ast.ts](https://github.com/flowr-analysis/flowr/tree/main//src/documentation/wiki-normalized-ast.ts)' on 2026-02-01, 19:10:59 UTC presenting an overview of flowR's normalized ast (v2.8.15, using R v4.5.2). Please do not edit this file/wiki page directly._
 
 _flowR_ produces a normalized version of R's abstract syntax tree (AST),
 offering the following benefits:
@@ -51,7 +51,7 @@ x"])
 
 ```
 	
-(The analysis required _0.2 ms_ (including parsing with the [tree-sitter](https://github.com/flowr-analysis/flowr/wiki/Engines) engine) within the generation environment.)
+(The analysis required _3.9 ms_ (including parsing with the [tree-sitter](https://github.com/flowr-analysis/flowr/wiki/Engines) engine) within the generation environment.)
 
 
 
@@ -2919,6 +2919,11 @@ Please note, that they usually operate on the <a href="https://github.com/flowr-
 you to decide how you want to traverse multiple files with a <a href="https://github.com/flowr-analysis/flowr/tree/main//src/r-bridge/lang-4.x/ast/model/nodes/r-project.ts#L20"><code><span title="Holds a collection of files (started with the expression list)">RProject</span></code></a> in the AST (you can, for example, simplify flat-map over the files).
 The <a href="https://github.com/flowr-analysis/flowr/tree/main//src/r-bridge/lang-4.x/ast/model/nodes/r-project.ts#L20"><code><span title="Holds a collection of files (started with the expression list)">RProject</span></code></a> node cannot appear nested within other nodes, so you can safely assume that any child of a node is not an <a href="https://github.com/flowr-analysis/flowr/tree/main//src/r-bridge/lang-4.x/ast/model/nodes/r-project.ts#L20"><code><span title="Holds a collection of files (started with the expression list)">RProject</span></code></a>.
 
+
+> [!TIP]
+> When working with names and identifiers, consider using the utilities provided with the <a href="https://github.com/flowr-analysis/flowr/tree/main//src/dataflow/environments/identifier.ts#L37"><code><span title="Helper functions to work with identifiers . Use Identifier.matches to check if two identifiers match according to R's scoping rules!">Identifier</span></code></a> object.
+
+
 ### Visitors
 
 If you want a simple visitor which traverses the AST, the <a href="https://github.com/flowr-analysis/flowr/tree/main//src/r-bridge/lang-4.x/ast/model/processing/visitor.ts#L119"><code><span title="Visits all node ids within a tree given by a respective root node using a depth-first search with prefix order.">visitAst</span></code></a> function is a good starting point.
@@ -3209,7 +3214,7 @@ class MyMathFold<Info> extends DefaultNormalizedAstFold<number, Info> {
 <i>Defined at <a href="https://github.com/flowr-analysis/flowr/tree/main//src/documentation/wiki-normalized-ast.ts#L35">./src/documentation/wiki-normalized-ast.ts#L35</a></i>
 
 
-Now, we can use the <a href="https://github.com/flowr-analysis/flowr/tree/main//src/project/flowr-analyzer.ts#L144"><code><span title="Central class for conducting analyses with FlowR. Use the FlowrAnalyzerBuilder to create a new instance. If you want the original pattern of creating a pipeline and running all steps, you can still do this with FlowrAnalyzer#runFull . To inspect the context of the analyzer, use FlowrAnalyzer#inspectContext (if you are a plugin and need to modify it, use FlowrAnalyzer#context instead).">FlowrAnalyzer</span></code></a> (see the [Analyzer](https://github.com/flowr-analysis/flowr/wiki/analyzer) wiki page) to get the [normalized AST](https://github.com/flowr-analysis/flowr/wiki/normalized-ast) and apply the fold:
+Now, we can use the <a href="https://github.com/flowr-analysis/flowr/tree/main//src/project/flowr-analyzer.ts#L144"><code><span title="Central class for conducting analyses with FlowR. Use the FlowrAnalyzerBuilder to create a new instance. If you want the original pattern of creating a pipeline and running all steps, you can still do this with FlowrAnalyzer#runFull . To inspect the context of the analyzer, use FlowrAnalyzer#inspectContext (if you are a plugin and need to modify it, use FlowrAnalyzer#context instead).">FlowrAnalyzer</span></code></a> (see the [Analyzer](https://github.com/flowr-analysis/flowr/wiki/wiki/analyzer) wiki page) to get the [normalized AST](https://github.com/flowr-analysis/flowr/wiki/wiki/normalized-ast) and apply the fold:
  
 
 ```ts
