+++ wiki/Overview.md
@@ -1,6 +1,6 @@
-_This document was generated from '[src/documentation/wiki-overview.ts](https://github.com/flowr-analysis/flowr/tree/main//src/documentation/wiki-overview.ts)' on 2026-02-01, 18:35:50 UTC presenting an overview of flowR's modules (v2.8.15, using R v4.5.2). Please do not edit this file/wiki page directly._
+_This document was generated from '[src/documentation/wiki-overview.ts](https://github.com/flowr-analysis/flowr/tree/main//src/documentation/wiki-overview.ts)' on 2026-02-01, 18:58:57 UTC presenting an overview of flowR's modules (v2.8.15, using R v4.5.2). Please do not edit this file/wiki page directly._
 First of all, if you have never used _flowR_ before,
-please refer to the [Setup](https://github.com/flowr-analysis/flowr/wiki/setup) wiki page first, 
+please refer to the [Setup](https://github.com/flowr-analysis/flowr/wiki/wiki/setup) wiki page first, 
 for instructions on how to install _flowR_.
 
 <!-- TOC -->
@@ -32,7 +32,7 @@ It is available with the [`benchmark`](#benchmark-the-slicer) script.
 The statistics module is mostly independent of the slicer and can be used to analyze R files regarding their use of function definitions, assignments, and more.  It is used to identify common patterns in R code and is available with the [`statistics`](#generate-usage-statistics-of-r-code) script.
 
 The [core](https://github.com/flowr-analysis/flowr/tree/main/src/core) module contains _flowR_'s read-eval-print loop (REPL) and 
-_flowR_'s server. Furthermore, it contains the root definitions of how _flowR_ slices (see the [Interface](https://github.com/flowr-analysis/flowr/wiki/interface) wiki page for more information).
+_flowR_'s server. Furthermore, it contains the root definitions of how _flowR_ slices (see the [Interface](https://github.com/flowr-analysis/flowr/wiki/wiki/interface) wiki page for more information).
 
 The [utility](https://github.com/flowr-analysis/flowr/tree/main/src/util) module is of no further interest for the usage of _flowR_
 
@@ -52,14 +52,14 @@ Similarly, we offer an [Addin for RStudio](https://github.com/flowr-analysis/rst
 ⚒️ If you compile the _flowR_ sources yourself, you can access _flowR_ by the main script `npm run flowr` or in the development mode `npm run main-dev`.
 
 Independent of your way of launching *flowr*, we will write simply `flowr` for either (🐳️)&nbsp;`docker run -it --rm eagleoutice/flowr:latest` or (⚒️)&nbsp;`npm run flowr`.
-See the [Setup](https://github.com/flowr-analysis/flowr/wiki/setup) wiki page for more information on how to get _flowR_ running.
+See the [Setup](https://github.com/flowr-analysis/flowr/wiki/wiki/setup) wiki page for more information on how to get _flowR_ running.
 
 ### The Read-Eval-Print Loop (REPL)
 
 Once you launched _flowR_, you should see a small&nbsp;`R>` prompt. Use `:help` to receive instructions on how to use the REPL and what features are available (most prominently, you can access all [scripts](#calling-the-scripts-directly) simply by adding a colon before them).
 In general, all commands start with a colon (`:`), everything else is interpreted as a&nbsp;R expression which is directly evaluated by the underlying R shell 
-(however, due to security concerns, you need to start _flowR_ with <span title="Description (Command Line Argument): Allow to access the underlying R session when using flowR (security warning: this allows the execution of arbitrary R code!)">`--r-session-access`</span> and use the `r-shell` [engine](https://github.com/flowr-analysis/flowr/wiki/engines) to allow this).
-See the [Interface](https://github.com/flowr-analysis/flowr/wiki/interface) wiki page for more information on usage and the available commands.
+(however, due to security concerns, you need to start _flowR_ with <span title="Description (Command Line Argument): Allow to access the underlying R session when using flowR (security warning: this allows the execution of arbitrary R code!)">`--r-session-access`</span> and use the `r-shell` [engine](https://github.com/flowr-analysis/flowr/wiki/wiki/engines) to allow this).
+See the [Interface](https://github.com/flowr-analysis/flowr/wiki/wiki/interface) wiki page for more information on usage and the available commands.
 The following GIF showcases a simple example session:
 
 ![Example of a simple REPL session](gif/repl-demo-opt.gif)
@@ -94,7 +94,7 @@ Sleep 20000ms
 ### The Server
 
 Instead of the REPL, you can start _flowR_ in "([TCP](https://de.wikipedia.org/wiki/Transmission_Control_Protocol)) server-mode" using `flowr --server` (write `flowr --help` to find out more). Together with the server option, you can configure the port with <span title="Description (Command Line Argument): The port to listen on, if --server is given.">`--port`</span>.
-The supported requests are documented alongside the internal documentation, see the [Interface](https://github.com/flowr-analysis/flowr/wiki/interface) wiki page for more information.
+The supported requests are documented alongside the internal documentation, see the [Interface](https://github.com/flowr-analysis/flowr/wiki/wiki/interface) wiki page for more information.
 
 <details>
     <summary>Small demonstration using netcat</summary>
@@ -137,7 +137,7 @@ Sleep 200ms
 </details>
 
 The server allows accessing the REPL as well
-(see the [Interface](https://github.com/flowr-analysis/flowr/wiki/interface) wiki page for more information).
+(see the [Interface](https://github.com/flowr-analysis/flowr/wiki/wiki/interface) wiki page for more information).
 
 ## Calling the Scripts Directly
