+++ wiki/Setup.md
@@ -1,4 +1,4 @@
-_This document was generated from '[src/documentation/wiki-setup.ts](https://github.com/flowr-analysis/flowr/tree/main//src/documentation/wiki-setup.ts)' on 2025-12-13, 00:07:27 UTC presenting an overview of flowR's setup instructions (v2.7.0, using R v4.5.2). Please do not edit this file/wiki page directly._
+_This document was generated from '[src/documentation/wiki-setup.ts](https://github.com/flowr-analysis/flowr/tree/main//src/documentation/wiki-setup.ts)' on 2026-02-01, 18:58:57 UTC presenting an overview of flowR's setup instructions (v2.8.15, using R v4.5.2). Please do not edit this file/wiki page directly._
 There are several ways to use _flowR_.
 You can download and build it from source, download the accompanying [docker image](https://hub.docker.com/r/eagleoutice/flowr), 
 or use its [Visual Studio Code extension](https://marketplace.visualstudio.com/items?itemName=code-inspect.vscode-flowr) and [RStudio Addin](https://github.com/flowr-analysis/rstudio-addin-flowr).
@@ -48,7 +48,7 @@ To start flowr as a server, you can run:
 docker run -it --rm -p1042:1042 eagleoutice/flowr --server
 ```
 
-For more information, see the [Interface](https://github.com/flowr-analysis/flowr/wiki/interface) wiki page.
+For more information, see the [Interface](https://github.com/flowr-analysis/flowr/wiki/wiki/interface) wiki page.
 
 ## ⚒️ Building From Scratch
